M: ios.codevn.net

+++ filter/build.sh
@@ -33,15 +33,58 @@ cat src/abpvn_title.tmp src/abpvn_general.txt src/abpvn_ad_domain.txt src/abpvn_
 sed -e '/^$/d' abpvn.tmp > abpvn.txt
 # abpvn_ublock.txt
 cat src/abpvn_title.tmp src/abpvn_general.txt src/abpvn_ad_domain.txt src/abpvn_elemhide.txt src/abpvn_whitelist.txt src/abpvn_whitelist_elemhide.txt src/abpvn_adult.txt src/abpvn_adult_elemhide.txt src/abpvn_ublock_specific.txt > abpvn_ublock.tmp
-sed -e '/^$/d' abpvn_ublock.tmp > abpvn_ublock.txt
+# Search useless rule by ublock specific
+function get_replace_rule() {
+    # $1 is input file
+    local INPUT_FILE="$1.tmp"
+    local REMOVE_RULE_TMP_FILE="$1_removal.tmp"
+    grep -E '^-(.*)\n' $INPUT_FILE > $REMOVE_RULE_TMP_FILE
+    local REPLACE_RULE=''
+    for removeRule in $(cat $REMOVE_RULE_TMP_FILE | grep -v '\.$'); do
+            printf -v removeRule "%q\n" $removeRule
+            THIS_REPLACE_RULE="s|$(echo $removeRule | cut -c 1-${#removeRule})||g;s|$(echo $removeRule | cut -c 2-${#removeRule})||g"
+            if [ "$REPLACE_RULE" == "" ]
+            then
+                REPLACE_RULE="$THIS_REPLACE_RULE"
+            else
+                REPLACE_RULE="$REPLACE_RULE;$THIS_REPLACE_RULE"
+            fi
+    done
+    echo $REPLACE_RULE
+}
+UBLOCK_REPLACE_RULE=$(get_replace_rule abpvn_ublock)
+if [ "$UBLOCK_REPLACE_RULE" == "" ]
+then
+    sed -e '/^$/d' abpvn_ublock.tmp > abpvn_ublock.txt
+else
+    echo "UBLOCK_REPLACE_RULE: $UBLOCK_REPLACE_RULE"
+    sed $UBLOCK_REPLACE_RULE abpvn_ublock.tmp > abpvn_ublock1.tmp 
+    sed -e '/^$/d' abpvn_ublock1.tmp > abpvn_ublock.txt
+fi
 # abpvn_adguard.txt
 cat src/abpvn_title.tmp src/abpvn_general.txt src/abpvn_ad_domain.txt src/abpvn_elemhide.txt src/abpvn_whitelist.txt src/abpvn_whitelist_elemhide.txt src/abpvn_adult.txt src/abpvn_adult_elemhide.txt src/abpvn_adguard_specific.txt > abpvn_adguard.tmp
-sed -e '/^$/d' abpvn_adguard.tmp > abpvn_adguard.txt
+ADGUARD_REPLACE_RULE=$(get_replace_rule abpvn_adguard)
+if [ "$ADGUARD_REPLACE_RULE" == "" ]
+then
+    sed -e '/^$/d' abpvn_adguard.tmp > abpvn_adguard.txt
+else
+    echo "ADGUARD_REPLACE_RULE: $ADGUARD_REPLACE_RULE"
+    sed $ADGUARD_REPLACE_RULE abpvn_adguard.tmp > abpvn_adguard1.tmp 
+    sed -e '/^$/d' abpvn_adguard1.tmp > abpvn_adguard.txt
+fi
 # abpvn_noelemhide.txt
 cat src/abpvn_title.tmp src/abpvn_general.txt src/abpvn_ad_domain.txt src/abpvn_whitelist.txt src/abpvn_adult.txt > abpvn_noelemhide.tmp
 sed -e '/^$/d' abpvn_noelemhide.tmp > abpvn_noelemhide.txt
 # abpvn_content_blocker.txt
 cat src/abpvn_title.tmp src/abpvn_general.txt src/abpvn_ad_domain.txt src/abpvn_elemhide.txt src/abpvn_whitelist.txt src/abpvn_whitelist_elemhide.txt src/abpvn_adult.txt src/abpvn_adult_elemhide.txt src/abpvn_content_blocker_hot_fix.txt > abpvn_content_blocker.tmp
-sed -e '/^$/d' abpvn_content_blocker.tmp > abpvn_content_blocker.txt
+CONTENT_BLOCKER_REPLACE_RULE=$(get_replace_rule abpvn_content_blocker)
+if [ "$CONTENT_BLOCKER_REPLACE_RULE" == "" ]
+then
+    sed -e '/^$/d' abpvn_content_blocker.tmp > abpvn_content_blocker.txt
+else
+    echo "CONTENT_BLOCKER_REPLACE_RULE: $CONTENT_BLOCKER_REPLACE_RULE"
+    sed $CONTENT_BLOCKER_REPLACE_RULE abpvn_content_blocker.tmp > abpvn_content_blocker1.tmp
+    sed -e '/^$/d' abpvn_content_blocker1.tmp > abpvn_content_blocker.txt
+fi
 # remove tmp file
 rm -rf *.tmp src/*.tmp
